Clean up deprecation
