--- src/blocks/block-cta/deprecated/1.5.2/attributes.js
@@ -1,89 +1,0 @@
-// Version 1_5_2 attributes
-
-export const callToAction_1_5_2_attr = {
-	buttonText: {
-		type: 'string',
-	},
-	buttonUrl: {
-		type: 'string',
-		source: 'attribute',
-		selector: 'a',
-		attribute: 'href',
-	},
-	buttonAlignment: {
-		type: 'string',
-		default: 'center'
-	},
-	buttonBackgroundColor: {
-		type: 'string',
-		default: '#3373dc'
-	},
-	buttonTextColor: {
-		type: 'string',
-		default: '#ffffff'
-	},
-	buttonSize: {
-		type: 'string',
-		default: 'ab-button-size-medium'
-	},
-	buttonShape: {
-		type: 'string',
-		default: 'ab-button-shape-rounded'
-	},
-	buttonTarget: {
-		type: 'boolean',
-		default: false
-	},
-	ctaTitle: {
-		type: 'array',
-		selector: '.ab-cta-title',
-		source: 'children',
-	},
-	titleFontSize: {
-		type: 'number',
-		default: '32',
-	},
-	ctaTextFontSize: {
-		type: 'number',
-	},
-	ctaText: {
-		type: 'array',
-		selector: '.ab-cta-text',
-		source: 'children',
-	},
-	ctaWidth: {
-		type: 'string',
-	},
-	ctaBackgroundColor: {
-		type: 'string',
-	},
-	ctaTextColor: {
-		type: 'string',
-		default: '#32373c'
-	},
-	imgURL: {
-		type: 'string',
-		source: 'attribute',
-		attribute: 'src',
-		selector: 'img',
-	},
-	imgID: {
-		type: 'number',
-	},
-	imgAlt: {
-		type: 'string',
-		source: 'attribute',
-		attribute: 'alt',
-		selector: 'img',
-	},
-	dimRatio: {
-		type: 'number',
-		default: 50,
-	},
-
-	// Deprecated
-	ctaTitleFontSize: {
-		type: 'string',
-		default: '32'
-	},
-}
